Refactor protobuf spec into Rust. WIP: Need to replace usages.

+++ nomos-da/network/core/src/lib.rs
@@ -6,5 +6,5 @@ pub mod swarm;
 #[cfg(test)]
 pub mod test_utils;
 
-pub type SubnetworkId = u32;
+pub type SubnetworkId = u16;
 pub use libp2p::PeerId;
